refactor(squid): rename providers to services
