+++ packages/adapter-squid/src/services/EvmChainService.ts
