+++ packages/adapter-squid/src/services/IEvmChainService.ts
