+++ packages/adapter-squid/src/services/ISquidProvider.ts
