+++ packages/adapter-squid/src/services/SquidProvider.ts
