+++ packages/adapter-squid/src/services/index.ts
